[feature] route get/timeline
