+++ package.json
@@ -20,6 +20,7 @@
     "joi": "^17.6.0",
     "jsonwebtoken": "^8.5.1",
     "nodemon": "^2.0.19",
-    "pg": "^8.7.3"
+    "pg": "^8.7.3",
+    "url-metadata": "^2.5.0"
   }
 }
