+++ src/utilities/getPostMetadata.js
@@ -1,0 +1,7 @@
+import urlMetadata from "url-metadata";
+
+function getPostMetadata(postUrl){
+    return urlMetadata(postUrl);
+}
+
+export {getPostMetadata}
